requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn
pycodestyle==2.10.0
PyJWT==2.6.0
psycopg2-binary
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
